feat(orga): moved shared participations results use cases and their tests
