+++ api/tests/prescription/campaign-participation/unit/application/jobs/participation-shared-job-controller_test.js
@@ -1,5 +1,5 @@
-import { usecases } from '../../../../../../lib/domain/usecases/index.js';
 import { ParticipationSharedJobController } from '../../../../../../src/prescription/campaign-participation/application/jobs/participation-shared-job-controller.js';
+import { usecases } from '../../../../../../src/prescription/campaign-participation/domain/usecases/index.js';
 import { expect, sinon } from '../../../../../test-helper.js';
 
 describe('Unit | Prescription | Application | Jobs | participationSharedJobController', function () {
